Updated FVector.

+++ GS_Engine/src/Game Studio/Id.cpp
@@ -8,11 +8,11 @@ Id::Id(const char * Text): HashedString(HashString(Text))
 
 uint32 Id::HashString(const char * Text)
 {
-	const uint16 Length = String::StringLength(Text);
+	const uint32 Length = String::StringLength(Text);
 
 	uint32 Hash = 0;
 
-	for (uint16 i = 0; i < Length; i++)
+	for (uint32 i = 0; i < Length; i++)
 	{
 		Hash += Text[i] * i * 33;
 	}
